+++ changelog.md
@@ -4,11 +4,12 @@
 
 - Disabled the enhanced editor toggle button by default.
 - Disabled the line break indicator by default.
-- Added an option to disable saving when exiting using the 'Escape' key.
 - Added a 'Revert' GUI button.
+- Added an option to disable saving when exiting using the 'Escape' key.
 - Added an option to conditionally prevent opening the sign editor.
 - Added an option to block movement key input until key release when opening the sign editor.
 - Added an option to automatically add text to placed signs.
+- Added an option to click through signs to most block entities.
 
 ## 1.2.0